Minor review changes and pom version update
