+++ addons/events/knative-eventing-addon/pom.xml
@@ -6,7 +6,7 @@
     <parent>
         <groupId>org.kie.kogito</groupId>
         <artifactId>events</artifactId>
-        <version>8.0.0-SNAPSHOT</version>
+        <version>1.0.0-SNAPSHOT</version>
     </parent>
     <artifactId>knative-eventing-addon</artifactId>
     <name>Kogito :: Add-Ons :: Events :: Knative Eventing</name>
