+++ addons/events/knative-eventing-addon/src/main/java/org/kie/kogito/events/knative/ce/decorators/CloudEventHttpOutgoingDecorator.java
@@ -1,3 +1,17 @@
+/*
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.kie.kogito.events.knative.ce.decorators;
 
 import io.smallrye.reactive.messaging.http.HttpResponseMetadata;
@@ -11,7 +25,7 @@ import java.util.Collections;
 /**
  * Decorators for Http CloudEvents outgoing messages
  */
-public final class CloudEventHttpOutgoingDecorator implements Decorator {
+public final class CloudEventHttpOutgoingDecorator implements MessageDecorator {
 
     /**
      * Metadata to include content-type for structured CloudEvents messages
