+++ addons/events/knative-eventing-addon/src/main/java/org/kie/kogito/events/knative/ce/decorators/MessageDecorator.java
@@ -12,20 +12,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito.events.knative.ce.http;
+package org.kie.kogito.events.knative.ce.decorators;
 
-import javax.ws.rs.core.MediaType;
+import org.eclipse.microprofile.reactive.messaging.Message;
 
 /**
- * Extends {@link MediaType} to CloudEvents support
+ * {@link MessageDecorator}s decorates the {@link Message} envelope with metadata and additional information in a given context.
  */
-// this shouldn't be provided by the CE SDK? Send a PR.
-public final class ExtMediaType {
+public interface MessageDecorator {
 
-    public static final String CLOUDEVENTS_JSON = "application/cloudevents+json";
-    public static final MediaType CLOUDEVENTS_JSON_TYPE = new MediaType("application", "cloudevents+json");
+    /**
+     * Decorates the given reactive message
+     *
+     * @param payload payload to decorate
+     * @param <T>     payload type
+     * @return payload in Message format decorated
+     */
+    <T> Message<T> decorate(T payload);
 
-    private ExtMediaType() {
-
-    }
 }
