+++ addons/events/knative-eventing-addon/src/test/java/org/kie/kogito/events/knative/ce/CloudEventConverterTest.java
@@ -1,19 +1,32 @@
+/*
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.kie.kogito.events.knative.ce;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.util.StdDateFormat;
-import io.cloudevents.CloudEvent;
-import io.cloudevents.core.builder.CloudEventBuilder;
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.kie.kogito.services.event.AbstractProcessDataEvent;
 
 import java.io.IOException;
 import java.net.URI;
 import java.util.TimeZone;
 import java.util.UUID;
 
-import static org.assertj.core.api.Assertions.*;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.util.StdDateFormat;
+import io.cloudevents.CloudEvent;
+import io.cloudevents.core.builder.CloudEventBuilder;
+import org.junit.jupiter.api.Test;
+import org.kie.kogito.services.event.AbstractProcessDataEvent;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 class CloudEventConverterTest {
 
@@ -100,6 +113,7 @@ class CloudEventConverterTest {
     }
 
     public static class PersonDataEvent extends AbstractProcessDataEvent<Person> {
+
         private String kogitoStartFromNode;
 
         public PersonDataEvent(String source, Person body, String kogitoProcessinstanceId, String kogitoParentProcessinstanceId, String kogitoRootProcessinstanceId, String kogitoProcessId, String kogitoRootProcessId, String kogitoProcessinstanceState, String kogitoAddons) {
@@ -120,6 +134,7 @@ class CloudEventConverterTest {
     }
 
     public static class Person {
+
         private String firstName;
         private String lastName;
         private String email;
