+++ addons/events/knative-eventing-addon/src/test/resources/application.properties
@@ -1,2 +1,1 @@
 quarkus.log.level=INFO
-quarkus.log.category."org.kie".level=INFO
